+++ src/cc/lib/distributed/client.h
@@ -56,27 +56,29 @@ class GRPCClient final
     void NeighborCount(std::span<const NodeId> node_ids, std::span<const Type> edge_types,
                        std::span<const snark::Timestamp> timestamps, std::span<uint64_t> output_neighbor_counts);
 
-    void FullNeighbor(std::span<const NodeId> node_ids, std::span<const Type> edge_types,
+    void FullNeighbor(bool return_edge_created_ts, std::span<const NodeId> node_ids, std::span<const Type> edge_types,
                       std::span<const snark::Timestamp> timestamps, std::vector<NodeId> &output_nodes,
                       std::vector<Type> &output_types, std::vector<float> &output_weights,
-                      std::span<uint64_t> output_neighbor_counts);
+                      std::vector<Timestamp> &out_edge_created_ts, std::span<uint64_t> output_neighbor_counts);
 
-    void WeightedSampleNeighbor(int64_t seed, std::span<const NodeId> node_ids, std::span<const Type> edge_types,
-                                std::span<const snark::Timestamp> timestamps, size_t count,
-                                std::span<NodeId> output_nodes, std::span<Type> output_types,
-                                std::span<float> output_weights, NodeId default_node_id, float default_weight,
-                                Type default_edge_type);
+    void WeightedSampleNeighbor(bool return_edge_created_ts, int64_t seed, std::span<const NodeId> node_ids,
+                                std::span<const Type> edge_types, std::span<const snark::Timestamp> timestamps,
+                                size_t count, std::span<NodeId> output_nodes, std::span<Type> output_types,
+                                std::span<float> output_weights, std::span<Timestamp> output_edge_created_ts,
+                                NodeId default_node_id, float default_weight, Type default_edge_type);
 
-    void UniformSampleNeighbor(bool without_replacement, int64_t seed, std::span<const NodeId> node_ids,
-                               std::span<const Type> edge_types, std::span<const snark::Timestamp> timestamps,
-                               size_t count, std::span<NodeId> output_nodes, std::span<Type> output_types,
-                               NodeId default_node_id, Type default_type);
+    void UniformSampleNeighbor(bool without_replacement, bool return_edge_created_ts, int64_t seed,
+                               std::span<const NodeId> node_ids, std::span<const Type> edge_types,
+                               std::span<const snark::Timestamp> timestamps, size_t count,
+                               std::span<NodeId> output_nodes, std::span<Type> output_types,
+                               std::span<Timestamp> output_edge_created_ts, NodeId default_node_id, Type default_type);
 
-    void LastNCreated(std::span<const NodeId> input_node_ids, std::span<Type> input_edge_types,
-                      std::span<const Timestamp> input_timestamps, size_t count, std::span<NodeId> output_neighbor_ids,
-                      std::span<Type> output_neighbor_types, std::span<float> neighbors_weights,
-                      std::span<Timestamp> output_timestamps, NodeId default_node_id, float default_weight,
-                      Type default_edge_type, Timestamp default_timestamp);
+    void LastNCreated(bool return_edge_created_ts, std::span<const NodeId> input_node_ids,
+                      std::span<Type> input_edge_types, std::span<const Timestamp> input_timestamps, size_t count,
+                      std::span<NodeId> output_neighbor_ids, std::span<Type> output_neighbor_types,
+                      std::span<float> neighbors_weights, std::span<Timestamp> output_timestamps,
+                      NodeId default_node_id, float default_weight, Type default_edge_type,
+                      Timestamp default_timestamp);
 
     uint64_t CreateSampler(bool is_edge, CreateSamplerRequest_Category category, std::span<Type> types);