+++ src/cc/tests/distributed_test.cc
@@ -528,12 +528,14 @@ TEST(DistributedTest, SampleNeighborsSingleServer)
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<float> output_weights(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    client.WeightedSampleNeighbor(21, std::span(input_nodes), std::span(input_types), {}, nb_count,
-                                  std::span(output_nodes), std::span(output_types), std::span(output_weights), -1, 0.0f,
-                                  -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -6);
+    client.WeightedSampleNeighbor(true, 21, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                                  std::span(output_nodes), std::span(output_types), std::span(output_weights),
+                                  std::span(output_edge_ts), -1, 0.0f, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>(6, 0));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({3, 3, 3, 4, 5, 5}));
     EXPECT_EQ(output_weights, std::vector<float>({2, 2, 2, 2, 2, 2}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, UniformSampleNeighborsSingleServer)
@@ -545,10 +547,12 @@ TEST(DistributedTest, UniformSampleNeighborsSingleServer)
     const size_t nb_count = 2;
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    client.UniformSampleNeighbor(false, 21, std::span(input_nodes), std::span(input_types), {}, nb_count,
-                                 std::span(output_nodes), std::span(output_types), -1, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -5);
+    client.UniformSampleNeighbor(false, true, 21, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                                 std::span(output_nodes), std::span(output_types), std::span(output_edge_ts), -1, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 0, 0, 0, 0, 0}));
-    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({3, 3, 3, 5, 5, 6}));
+    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({3, 3, 5, 2, 4, 4}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, UniformSampleNeighborsWithoutReplacementSingleServer)
@@ -560,10 +564,12 @@ TEST(DistributedTest, UniformSampleNeighborsWithoutReplacementSingleServer)
     const size_t nb_count = 2;
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    client.UniformSampleNeighbor(true, 21, std::span(input_nodes), std::span(input_types), {}, nb_count,
-                                 std::span(output_nodes), std::span(output_types), -1, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -2);
+    client.UniformSampleNeighbor(true, true, 21, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                                 std::span(output_nodes), std::span(output_types), std::span(output_edge_ts), -1, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 0, 0, 0, 0, 0}));
-    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({1, 3, 2, 4, 6, 3}));
+    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({1, 4, 5, 4, 3, 5}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 using ServerList = std::vector<std::shared_ptr<snark::GRPCServer>>;
@@ -613,11 +619,14 @@ TEST(DistributedTest, SampleNeighborsMultipleServers)
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<float> output_weights(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    c.WeightedSampleNeighbor(23, std::span(input_nodes), std::span(input_types), {}, nb_count, std::span(output_nodes),
-                             std::span(output_types), std::span(output_weights), -1, 0.0f, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -4);
+    c.WeightedSampleNeighbor(true, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                             std::span(output_nodes), std::span(output_types), std::span(output_weights),
+                             std::span(output_edge_ts), -1, 0.0f, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>(6, 0));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({2, 2, 57, 56, 80, 81}));
     EXPECT_EQ(output_weights, std::vector<float>({2, 2, 2, 1, 1, 2}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, SampleNeighborsMultipleServersMissingNeighbors)
@@ -631,11 +640,14 @@ TEST(DistributedTest, SampleNeighborsMultipleServersMissingNeighbors)
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<float> output_weights(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    c.WeightedSampleNeighbor(23, std::span(input_nodes), std::span(input_types), {}, nb_count, std::span(output_nodes),
-                             std::span(output_types), std::span(output_weights), -1, 0.0f, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -2);
+    c.WeightedSampleNeighbor(true, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                             std::span(output_nodes), std::span(output_types), std::span(output_weights),
+                             std::span(output_edge_ts), -1, 0.0f, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>(6, -1));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>(6, -1));
     EXPECT_EQ(output_weights, std::vector<float>(6, 0));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, UniformSampleNeighborsMultipleServers)
@@ -648,10 +660,12 @@ TEST(DistributedTest, UniformSampleNeighborsMultipleServers)
     const size_t nb_count = 2;
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    c.UniformSampleNeighbor(false, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
-                            std::span(output_nodes), std::span(output_types), -1, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -3);
+    c.UniformSampleNeighbor(false, true, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                            std::span(output_nodes), std::span(output_types), std::span(output_edge_ts), -1, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 0, 0, 0, 0, 0}));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({2, 2, 57, 56, 80, 81}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, UniformSampleNeighborsWithoutReplacementMultipleServers)
@@ -664,10 +678,12 @@ TEST(DistributedTest, UniformSampleNeighborsWithoutReplacementMultipleServers)
     const size_t nb_count = 2;
     std::vector<snark::NodeId> output_nodes(nb_count * input_nodes.size());
     std::vector<snark::Type> output_types(nb_count * input_nodes.size(), -1);
-    c.UniformSampleNeighbor(true, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
-                            std::span(output_nodes), std::span(output_types), -1, -1);
+    std::vector<snark::Timestamp> output_edge_ts(nb_count * input_nodes.size(), -2);
+    c.UniformSampleNeighbor(true, true, 23, std::span(input_nodes), std::span(input_types), {}, nb_count,
+                            std::span(output_nodes), std::span(output_types), std::span(output_edge_ts), -1, -1);
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 0, 0, 0, 0, 0}));
-    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({2, 4, 59, 57, 81, 79}));
+    EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({3, 2, 58, 57, 80, 79}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(6, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, NeighborCountMultipleServers)
@@ -864,12 +880,14 @@ TEST(DistributedTest, FullNeighborsMultipleTypesMultipleServers)
     std::vector<snark::Type> output_types;
     std::vector<float> output_weights;
     std::vector<uint64_t> output_counts(input_nodes.size());
-    c.FullNeighbor(std::span(input_nodes), std::span(input_types), {}, output_nodes, output_types, output_weights,
-                   std::span(output_counts));
+    std::vector<snark::Timestamp> output_edge_ts;
+    c.FullNeighbor(true, std::span(input_nodes), std::span(input_types), {}, output_nodes, output_types, output_weights,
+                   output_edge_ts, std::span(output_counts));
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 1, 1}));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({1, 2, 3}));
     EXPECT_EQ(output_weights, std::vector<float>({1, 2, 1}));
     EXPECT_EQ(output_counts, std::vector<uint64_t>({3}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(3, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 TEST(DistributedTest, FullNeighborsMultipleServers)
@@ -883,12 +901,14 @@ TEST(DistributedTest, FullNeighborsMultipleServers)
     std::vector<snark::Type> output_types;
     std::vector<float> output_weights;
     std::vector<uint64_t> output_counts(input_nodes.size());
-    c.FullNeighbor(std::span(input_nodes), std::span(input_types), {}, output_nodes, output_types, output_weights,
-                   std::span(output_counts));
+    std::vector<snark::Timestamp> output_edge_ts;
+    c.FullNeighbor(true, std::span(input_nodes), std::span(input_types), {}, output_nodes, output_types, output_weights,
+                   output_edge_ts, std::span(output_counts));
     EXPECT_EQ(output_types, std::vector<snark::Type>({0, 0, 0, 0, 0, 0, 0, 0}));
     EXPECT_EQ(output_nodes, std::vector<snark::NodeId>({1, 2, 3, 4, 56, 57, 58, 59}));
     EXPECT_EQ(output_weights, std::vector<float>({1, 2, 1, 2, 1, 2, 1, 2}));
     EXPECT_EQ(output_counts, std::vector<uint64_t>({4, 4}));
+    EXPECT_EQ(output_edge_ts, std::vector<snark::Timestamp>(8, snark::PLACEHOLDER_TIMESTAMP));
 }
 
 std::pair<ServerList, std::shared_ptr<snark::GRPCClient>> CreateMultiServerSplitFeaturesEnvironment(