+++ src/cc/tests/temporal_test.cc
@@ -191,17 +191,19 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     std::vector<snark::NodeId> output_neighbor_ids;
     std::vector<snark::Type> output_neighbor_types;
     std::vector<float> output_neighbors_weights;
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    std::vector<snark::Timestamp> output_edge_created_ts;
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_TRUE(output_neighbor_ids.empty());
     EXPECT_TRUE(output_neighbor_types.empty());
     EXPECT_TRUE(output_neighbors_weights.empty());
+    EXPECT_TRUE(output_edge_created_ts.empty());
     EXPECT_EQ(std::vector<uint64_t>({0, 0}), output_neighbors_count);
 
     ts = {0, 0};
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({1, 2, 3}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -209,14 +211,16 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({1.f, 2.f, 1.f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({0, 0, 0}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({2, 1}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
     // Check for different singe edge type filter
     types = {1};
     ts = {2, 2};
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({5}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -224,13 +228,15 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({7.f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({2}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({0, 1}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
     // Check for both edge types
     types = {0, 1};
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({5}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -238,6 +244,8 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({7.f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({2}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({0, 1}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
@@ -245,12 +253,13 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     types = {-1, 100};
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_TRUE(output_neighbor_ids.empty());
     EXPECT_TRUE(output_neighbor_types.empty());
     EXPECT_TRUE(output_neighbors_weights.empty());
+    EXPECT_TRUE(output_edge_created_ts.empty());
     EXPECT_EQ(std::vector<uint64_t>({0, 0}), output_neighbors_count);
 
     // Invalid node ids
@@ -258,12 +267,13 @@ TEST_F(TemporalTest, GetFullNeighborSinglePartition)
     types = {0, 1};
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
-    m_single_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                           output_neighbor_types, output_neighbors_weights,
+    m_single_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                           output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                            std::span(output_neighbors_count));
     EXPECT_TRUE(output_neighbor_ids.empty());
     EXPECT_TRUE(output_neighbor_types.empty());
     EXPECT_TRUE(output_neighbors_weights.empty());
+    EXPECT_TRUE(output_edge_created_ts.empty());
     EXPECT_EQ(std::vector<uint64_t>({0, 0}), output_neighbors_count);
 }
 
@@ -312,8 +322,9 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     std::vector<snark::NodeId> output_neighbor_ids;
     std::vector<snark::Type> output_neighbor_types;
     std::vector<float> output_neighbors_weights;
-    m_multi_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                          output_neighbor_types, output_neighbors_weights,
+    std::vector<snark::Timestamp> output_edge_created_ts;
+    m_multi_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                          output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                           std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({5, 7}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -321,13 +332,15 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({1.f, 3.0f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({1, 2}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({0, 2}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
     // Check for multiple edge types
     types = {0, 1};
-    m_multi_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                          output_neighbor_types, output_neighbors_weights,
+    m_multi_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                          output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                           std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({4, 5, 7}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -335,14 +348,16 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({1.f, 1.f, 3.f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({1, 1, 2}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({0, 3}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
     // Check for different singe edge type filter
     types = {0};
     ts = {2, 2};
-    m_multi_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                          output_neighbor_types, output_neighbors_weights,
+    m_multi_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                          output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                           std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({4}), output_neighbor_ids);
     output_neighbor_ids.clear();
@@ -350,6 +365,8 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     output_neighbor_types.clear();
     EXPECT_EQ(std::vector<float>({1.f}), output_neighbors_weights);
     output_neighbors_weights.clear();
+    EXPECT_EQ(std::vector<snark::Timestamp>({1}), output_edge_created_ts);
+    output_edge_created_ts.clear();
     EXPECT_EQ(std::vector<uint64_t>({0, 1}), output_neighbors_count);
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
@@ -357,12 +374,13 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     types = {-1, 100};
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
-    m_multi_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                          output_neighbor_types, output_neighbors_weights,
+    m_multi_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                          output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                           std::span(output_neighbors_count));
     EXPECT_TRUE(output_neighbor_ids.empty());
     EXPECT_TRUE(output_neighbor_types.empty());
     EXPECT_TRUE(output_neighbors_weights.empty());
+    EXPECT_TRUE(output_edge_created_ts.empty());
     EXPECT_EQ(std::vector<uint64_t>({0, 0}), output_neighbors_count);
 
     // Invalid node ids
@@ -370,12 +388,13 @@ TEST_F(TemporalTest, GetFullNeighborMultiplePartitions)
     types = {0, 1};
     std::fill_n(output_neighbors_count.begin(), 2, -1);
 
-    m_multi_partition_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                          output_neighbor_types, output_neighbors_weights,
+    m_multi_partition_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                          output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                           std::span(output_neighbors_count));
     EXPECT_TRUE(output_neighbor_ids.empty());
     EXPECT_TRUE(output_neighbor_types.empty());
     EXPECT_TRUE(output_neighbors_weights.empty());
+    EXPECT_TRUE(output_edge_created_ts.empty());
     EXPECT_EQ(std::vector<uint64_t>({0, 0}), output_neighbors_count);
 }
 
@@ -391,10 +410,11 @@ TEST_F(TemporalTest, GetSampleNeighborsMultiplePartitions)
     std::vector<snark::Type> output_neighbor_types(sample_count * nodes.size());
     std::vector<float> output_neighbors_weights(sample_count * nodes.size());
     std::vector<float> output_neighbors_total_weights(nodes.size());
-    m_multi_partition_graph->SampleNeighbor(33, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                            std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                            std::span(output_neighbors_weights),
-                                            std::span(output_neighbors_total_weights), 42, 0.5f, 13);
+    std::vector<snark::Timestamp> output_edge_created_ts(sample_count * nodes.size(), -2);
+    m_multi_partition_graph->SampleNeighbor(
+        true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_weights),
+        std::span(output_neighbors_total_weights), std::span(output_edge_created_ts), 42, 0.5f, 13);
 
     // Only available neighbor based on time/type is 5
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 5, 5}), output_neighbor_ids);
@@ -405,13 +425,16 @@ TEST_F(TemporalTest, GetSampleNeighborsMultiplePartitions)
     std::fill(std::begin(output_neighbors_weights), std::end(output_neighbors_weights), -1);
     EXPECT_EQ(std::vector<float>({0.f, 4.0f}), output_neighbors_total_weights);
     std::fill(std::begin(output_neighbors_total_weights), std::end(output_neighbors_total_weights), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, 1}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
 
     // Check for multiple edge types
     types = {0, 1};
-    m_multi_partition_graph->SampleNeighbor(36, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                            std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                            std::span(output_neighbors_weights),
-                                            std::span(output_neighbors_total_weights), 42, 0.5f, 13);
+    m_multi_partition_graph->SampleNeighbor(
+        true, 36, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_weights),
+        std::span(output_neighbors_total_weights), std::span(output_edge_created_ts), 42, 0.5f, 13);
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 7, 5}), output_neighbor_ids);
     std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 1}), output_neighbor_types);
@@ -420,14 +443,17 @@ TEST_F(TemporalTest, GetSampleNeighborsMultiplePartitions)
     std::fill(std::begin(output_neighbors_weights), std::end(output_neighbors_weights), -1);
     EXPECT_EQ(std::vector<float>({0.f, 5.f}), output_neighbors_total_weights);
     std::fill(std::begin(output_neighbors_total_weights), std::end(output_neighbors_total_weights), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 2, 1}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
 
     // Check for different singe edge type filter
     types = {1};
     ts = {0, 0};
-    m_multi_partition_graph->SampleNeighbor(37, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                            std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                            std::span(output_neighbors_weights),
-                                            std::span(output_neighbors_total_weights), 42, 0.5f, 13);
+    m_multi_partition_graph->SampleNeighbor(
+        true, 37, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_weights),
+        std::span(output_neighbors_total_weights), std::span(output_edge_created_ts), 42, 0.5f, 13);
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 6, 6}), output_neighbor_ids);
     std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 1}), output_neighbor_types);
@@ -436,14 +462,17 @@ TEST_F(TemporalTest, GetSampleNeighborsMultiplePartitions)
     std::fill(std::begin(output_neighbors_weights), std::end(output_neighbors_weights), -1);
     EXPECT_EQ(std::vector<float>({0.f, 1.5f}), output_neighbors_total_weights);
     std::fill(std::begin(output_neighbors_total_weights), std::end(output_neighbors_total_weights), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 0, 0}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
 
     // Check returns 0 for unsatisfying edge types
     types = {-1, 100};
 
-    m_multi_partition_graph->SampleNeighbor(33, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                            std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                            std::span(output_neighbors_weights),
-                                            std::span(output_neighbors_total_weights), 42, 0.5f, 13);
+    m_multi_partition_graph->SampleNeighbor(
+        true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_weights),
+        std::span(output_neighbors_total_weights), std::span(output_edge_created_ts), 42, 0.5f, 13);
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 42, 42}), output_neighbor_ids);
     std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 13, 13}), output_neighbor_types);
@@ -452,20 +481,110 @@ TEST_F(TemporalTest, GetSampleNeighborsMultiplePartitions)
     std::fill(std::begin(output_neighbors_weights), std::end(output_neighbors_weights), -1);
     EXPECT_EQ(std::vector<float>({0.f, 0.f}), output_neighbors_total_weights);
     std::fill(std::begin(output_neighbors_total_weights), std::end(output_neighbors_total_weights), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>(4, snark::PLACEHOLDER_TIMESTAMP), output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
 
     // Invalid node ids
     nodes = {99, 100};
     types = {0, 1};
 
-    m_multi_partition_graph->SampleNeighbor(33, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                            std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                            std::span(output_neighbors_weights),
-                                            std::span(output_neighbors_total_weights), 42, 0.5f, 13);
+    m_multi_partition_graph->SampleNeighbor(
+        true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_weights),
+        std::span(output_neighbors_total_weights), std::span(output_edge_created_ts), 42, 0.5f, 13);
 
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 42, 42}), output_neighbor_ids);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 13, 13}), output_neighbor_types);
     EXPECT_EQ(std::vector<float>({0.5f, 0.5f, 0.5f, 0.5f}), output_neighbors_weights);
     EXPECT_EQ(std::vector<float>({0.f, 0.f}), output_neighbors_total_weights);
+    EXPECT_EQ(std::vector<snark::Timestamp>(4, snark::PLACEHOLDER_TIMESTAMP), output_edge_created_ts);
+}
+
+TEST_F(TemporalTest, GetUniformSampleNeighborMultiplePartitions)
+{
+    // Check for singe edge type filter
+    std::vector<snark::NodeId> nodes = {0, 1};
+    std::vector<snark::Type> types = {1};
+    std::vector<snark::Timestamp> ts = {2, 2};
+    size_t sample_count = 2;
+
+    std::vector<snark::NodeId> output_neighbor_ids(sample_count * nodes.size());
+    std::vector<snark::Type> output_neighbor_types(sample_count * nodes.size());
+    std::vector<uint64_t> output_neighbors_total_counts(nodes.size());
+    std::vector<snark::Timestamp> output_edge_created_ts(sample_count * nodes.size(), -2);
+    m_multi_partition_graph->UniformSampleNeighbor(
+        false, true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count,
+        std::span(output_neighbor_ids), std::span(output_neighbor_types), std::span(output_neighbors_total_counts),
+        std::span(output_edge_created_ts), 42, 13);
+
+    // Only available neighbors based on time/type are 5 and 7
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 7, 5}), output_neighbor_ids);
+    std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 1}), output_neighbor_types);
+    std::fill(std::begin(output_neighbor_types), std::end(output_neighbor_types), -1);
+    std::fill(std::begin(output_neighbors_total_counts), std::end(output_neighbors_total_counts), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 2, 1}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
+
+    m_multi_partition_graph->UniformSampleNeighbor(
+        true, true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_total_counts), std::span(output_edge_created_ts),
+        42, 13);
+
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 5, 7}), output_neighbor_ids);
+    std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 1}), output_neighbor_types);
+    std::fill(std::begin(output_neighbor_types), std::end(output_neighbor_types), -1);
+    EXPECT_EQ(std::vector<uint64_t>({0, 2}), output_neighbors_total_counts);
+    std::fill(std::begin(output_neighbors_total_counts), std::end(output_neighbors_total_counts), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, 2}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
+
+    ts = {1, 1};
+    m_multi_partition_graph->UniformSampleNeighbor(
+        false, true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count,
+        std::span(output_neighbor_ids), std::span(output_neighbor_types), std::span(output_neighbors_total_counts),
+        std::span(output_edge_created_ts), 42, 13);
+    // Only available neighbors based on time/type is 5
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 5, 5}), output_neighbor_ids);
+    std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 1}), output_neighbor_types);
+    std::fill(std::begin(output_neighbor_types), std::end(output_neighbor_types), -1);
+    EXPECT_EQ(std::vector<uint64_t>({0, 1}), output_neighbors_total_counts);
+    std::fill(std::begin(output_neighbors_total_counts), std::end(output_neighbors_total_counts), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, 1}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
+
+    m_multi_partition_graph->UniformSampleNeighbor(
+        true, true, 33, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_neighbors_total_counts), std::span(output_edge_created_ts),
+        42, 13);
+
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 5, 42}), output_neighbor_ids);
+    std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -1);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 1, 13}), output_neighbor_types);
+    std::fill(std::begin(output_neighbor_types), std::end(output_neighbor_types), -1);
+    EXPECT_EQ(std::vector<uint64_t>({0, 1}), output_neighbors_total_counts);
+    std::fill(std::begin(output_neighbors_total_counts), std::end(output_neighbors_total_counts), 0);
+    EXPECT_EQ(std::vector<snark::Timestamp>(
+                  {snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, snark::PLACEHOLDER_TIMESTAMP}),
+              output_edge_created_ts);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
+
+    // Check for multiple edge types
+    types = {0, 1};
+    ts = {0, 0};
+    m_multi_partition_graph->UniformSampleNeighbor(
+        false, true, 36, std::span(nodes), std::span(types), std::span(ts), sample_count,
+        std::span(output_neighbor_ids), std::span(output_neighbor_types), std::span(output_neighbors_total_counts),
+        std::span(output_edge_created_ts), 42, 13);
+    EXPECT_EQ(std::vector<snark::NodeId>({1, 2, 6, 3}), output_neighbor_ids);
+    EXPECT_EQ(std::vector<snark::Type>({0, 0, 1, 0}), output_neighbor_types);
+    EXPECT_EQ(std::vector<uint64_t>({2, 2}), output_neighbors_total_counts);
+    EXPECT_EQ(std::vector<snark::Timestamp>({0, 0, 0, 0}), output_edge_created_ts);
 }
 
 TEST_F(TemporalTest, GetFullNeighborDistributed)
@@ -479,13 +598,15 @@ TEST_F(TemporalTest, GetFullNeighborDistributed)
     std::vector<snark::NodeId> output_neighbor_ids;
     std::vector<snark::Type> output_neighbor_types;
     std::vector<float> output_neighbors_weights;
-    m_distributed_graph->FullNeighbor(std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
-                                      output_neighbor_types, output_neighbors_weights,
+    std::vector<snark::Timestamp> output_edge_created_ts;
+    m_distributed_graph->FullNeighbor(true, std::span(nodes), std::span(types), std::span(ts), output_neighbor_ids,
+                                      output_neighbor_types, output_neighbors_weights, output_edge_created_ts,
                                       std::span(output_neighbors_count));
     EXPECT_EQ(std::vector<snark::NodeId>({5, 7}), output_neighbor_ids);
     EXPECT_EQ(std::vector<snark::Type>({1, 1}), output_neighbor_types);
     EXPECT_EQ(std::vector<float>({1.f, 3.0f}), output_neighbors_weights);
     EXPECT_EQ(std::vector<uint64_t>({0, 2}), output_neighbors_count);
+    EXPECT_EQ(std::vector<snark::Timestamp>({1, 2}), output_edge_created_ts);
 }
 
 TEST_F(TemporalTest, GetNeighborCountDistributed)
@@ -515,12 +636,48 @@ TEST_F(TemporalTest, GetSampleNeighborsDistributed)
     std::vector<snark::NodeId> output_neighbor_ids(sample_count * nodes.size());
     std::vector<snark::Type> output_neighbor_types(sample_count * nodes.size());
     std::vector<float> output_neighbors_weights(sample_count * nodes.size());
-    m_distributed_graph->WeightedSampleNeighbor(37, std::span(nodes), std::span(types), std::span(ts), sample_count,
-                                                std::span(output_neighbor_ids), std::span(output_neighbor_types),
-                                                std::span(output_neighbors_weights), 42, 0.5f, 13);
+    std::vector<snark::Timestamp> output_edge_created_ts(sample_count * nodes.size(), -2);
+    m_distributed_graph->WeightedSampleNeighbor(true, 37, std::span(nodes), std::span(types), std::span(ts),
+                                                sample_count, std::span(output_neighbor_ids),
+                                                std::span(output_neighbor_types), std::span(output_neighbors_weights),
+                                                std::span(output_edge_created_ts), 42, 0.5f, 13);
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 4, 4}), output_neighbor_ids);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 0, 0}), output_neighbor_types);
     EXPECT_EQ(std::vector<float>({0.5f, 0.5f, 1.f, 1.f}), output_neighbors_weights);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, 1}),
+              output_edge_created_ts);
+}
+
+TEST_F(TemporalTest, GetUniformSampleNeighborsDistributed)
+{
+    std::vector<snark::NodeId> nodes = {0, 1};
+    // Keep types = 0 to keep test stable and avoid merging neighbors from multiple shards.
+    std::vector<snark::Type> types = {0};
+    std::vector<snark::Timestamp> ts = {2, 2};
+    size_t sample_count = 2;
+
+    std::vector<snark::NodeId> output_neighbor_ids(sample_count * nodes.size());
+    std::vector<snark::Type> output_neighbor_types(sample_count * nodes.size());
+    std::vector<snark::Timestamp> output_edge_created_ts(sample_count * nodes.size(), -2);
+    m_distributed_graph->UniformSampleNeighbor(
+        true, true, 37, std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
+        std::span(output_neighbor_types), std::span(output_edge_created_ts), 42, 13);
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 4, 42}), output_neighbor_ids);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 0, 13}), output_neighbor_types);
+    EXPECT_EQ(std::vector<snark::Timestamp>(
+                  {snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, snark::PLACEHOLDER_TIMESTAMP}),
+              output_edge_created_ts);
+
+    std::fill(std::begin(output_neighbor_ids), std::end(output_neighbor_ids), -2);
+    std::fill(std::begin(output_neighbor_types), std::end(output_neighbor_types), -2);
+    std::fill(std::begin(output_edge_created_ts), std::end(output_edge_created_ts), -2);
+    m_distributed_graph->UniformSampleNeighbor(
+        false, true, 37, std::span(nodes), std::span(types), std::span(ts), sample_count,
+        std::span(output_neighbor_ids), std::span(output_neighbor_types), std::span(output_edge_created_ts), 42, 13);
+    EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 4, 4}), output_neighbor_ids);
+    EXPECT_EQ(std::vector<snark::Type>({13, 13, 0, 0}), output_neighbor_types);
+    EXPECT_EQ(std::vector<snark::Timestamp>({snark::PLACEHOLDER_TIMESTAMP, snark::PLACEHOLDER_TIMESTAMP, 1, 1}),
+              output_edge_created_ts);
 }
 
 namespace
@@ -536,9 +693,9 @@ template <typename Graph> void GetLastNCreatedNeighbors(Graph &g)
     std::vector<snark::Type> output_neighbor_types(sample_count * nodes.size());
     std::vector<float> output_neighbors_weights(sample_count * nodes.size());
     std::vector<snark::Timestamp> output_timestamps(sample_count * nodes.size());
-    g->LastNCreated(std::span(nodes), std::span(types), std::span(ts), sample_count, std::span(output_neighbor_ids),
-                    std::span(output_neighbor_types), std::span(output_neighbors_weights), std::span(output_timestamps),
-                    42, 0.5f, 13, 99);
+    g->LastNCreated(true, std::span(nodes), std::span(types), std::span(ts), sample_count,
+                    std::span(output_neighbor_ids), std::span(output_neighbor_types),
+                    std::span(output_neighbors_weights), std::span(output_timestamps), 42, 0.5f, 13, 99);
     EXPECT_EQ(std::vector<snark::NodeId>({42, 42, 4, 42}), output_neighbor_ids);
     EXPECT_EQ(std::vector<snark::Type>({13, 13, 0, 13}), output_neighbor_types);
     EXPECT_EQ(std::vector<float>({0.5f, 0.5f, 1.f, 0.5f}), output_neighbors_weights);