+++ src/python/deepgnn/graph_engine/snark/local.py
@@ -107,7 +107,11 @@ class Client(Graph):
         alpha: float = 0.5,
         eps: float = 0.0001,
         timestamps: Union[List[int], np.ndarray] = None,
-    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+        return_edge_created_ts: bool = False,
+    ) -> Union[
+        Tuple[np.ndarray, np.ndarray, np.ndarray],
+        Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray],
+    ]:
         """Sample node neighbors."""
         if strategy == "byweight":
             result = self.graph.weighted_sample_neighbors(  # type: ignore
@@ -118,8 +122,16 @@ class Client(Graph):
                 default_weight,
                 seed=int(random.getrandbits(64)),
                 timestamps=timestamps,
+                return_edge_created_ts=return_edge_created_ts,
             )
-            return result[0], result[1], result[2], np.empty((1), dtype=np.int32)
+            if return_edge_created_ts:
+                return result
+            return (
+                result[0],
+                result[1],
+                result[2],
+                np.full(result[0].shape, -1, dtype=np.int64),
+            )
         without_replacement = strategy == "randomwithoutreplacement"
         if strategy in ["random", "randomwithoutreplacement"]:
             result = self.graph.uniform_sample_neighbors(  # type: ignore
@@ -131,12 +143,15 @@ class Client(Graph):
                 default_edge_type,
                 seed=int(random.getrandbits(64)),
                 timestamps=timestamps,
+                return_edge_created_ts=return_edge_created_ts,
             )
             return (
                 result[0],
                 np.empty(result[0].shape, dtype=np.float32),
                 result[1],
-                np.empty((1), dtype=np.int32),
+                result[2]
+                if return_edge_created_ts
+                else np.full(result[0].shape, -1, dtype=np.int64),
             )
         if strategy == "ppr-go":
             result = self.graph.ppr_neighbors(  # type: ignore
@@ -153,7 +168,7 @@ class Client(Graph):
                 result[0],
                 result[1],
                 np.empty(0, dtype=np.int32),
-                np.empty(0, dtype=np.int32),
+                np.full(result[0].shape, -1, dtype=np.int64),
             )
 
         if strategy == "lastn":
@@ -165,6 +180,7 @@ class Client(Graph):
                 default_node=default_node,
                 default_weight=default_weight,
                 default_edge_type=default_edge_type,
+                return_edge_created_ts=return_edge_created_ts,
             )
 
         raise NotImplementedError(f"Unknown strategy type {strategy}")
@@ -225,10 +241,17 @@ class Client(Graph):
         nodes: np.ndarray,
         edge_types: Union[int, np.ndarray],
         timestamps: Union[List[int], np.ndarray] = None,
-    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+        return_edge_created_ts: bool = False,
+    ) -> Union[
+        Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray],
+        Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray, np.ndarray],
+    ]:
         """Fetch full information about node neighbors."""
         return self.graph.neighbors(
-            nodes, self.__check_types(edge_types), timestamps=timestamps
+            nodes,
+            self.__check_types(edge_types),
+            timestamps=timestamps,
+            return_edge_created_ts=return_edge_created_ts,
         )
 
     def neighbor_count(