+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `return_edge_created_ts` argument to neighbor sampling methods to return timestamps when edges connecting nodes were created.
+
+### Fixed
+- Uniform sampling works in temporal graphs.
+
 ## [0.1.60] - 2022-04-18
 
 ### Added